requirements.txt
opencv-python
numpy
face_recognition
pillow
pytesseract
imutils
easyocr 
RPi.GPIO  